+++ examples/liveliness/src/liveliness.rs
@@ -31,8 +31,8 @@ fn delete_subscription(ctx: &ArcContext<AgentProps>, id: &str) -> Result<()> {
 
 #[tokio::main]
 async fn main() -> Result<()> {
-	// a tracing subscriber writing logs
-	tracing_subscriber::fmt::init();
+	// initialize tracing/logging
+	init_tracing();
 
 	// create & initialize agents properties
 	let properties = AgentProps { num: 1 };