+++ examples/pingpong/src/ping.rs
@@ -45,8 +45,8 @@ fn pong_received(_ctx: &ArcContext<AgentProps>, message: Message) -> Result<()> 
 
 #[tokio::main]
 async fn main() -> Result<()> {
-	// a tracing subscriber writing logs
-	tracing_subscriber::fmt::init();
+	// initialize tracing/logging
+	init_tracing();
 
 	// create & initialize agents properties
 	let properties = AgentProps { counter: 0 };