+++ examples/pingpong/src/pong.rs
@@ -38,8 +38,8 @@ fn ping_received(ctx: &ArcContext<AgentProps>, message: Message) -> Result<()> {
 
 #[tokio::main]
 async fn main() -> Result<()> {
-	// a tracing subscriber writing logs
-	tracing_subscriber::fmt::init();
+	// initialize tracing/logging
+	init_tracing();
 
 	// create & initialize agents properties
 	let properties = AgentProps {};